[#4]api test implement for Verification
