+++ package.json
@@ -25,7 +25,7 @@
     },
     "testRegex": "\\.(test|spec)\\.((js|ts))$",
     "setupFilesAfterEnv": [
-      "<rootDir>/src/__tests__/setup.ts"
+      "<rootDir>/src/testUtils/setup.ts"
     ]
   },
   "author": "naamoonoo",
