--- src/api/hello.graphql
@@ -1,3 +1,0 @@
-type Query {
-	hello: [Verification]
-}
