--- src/api/hello.resolvers.ts
@@ -1,9 +1,0 @@
-import { Verification } from "../entities/Verification";
-
-export default {
-	Query: {
-		hello: () => {
-			return Verification.find({});
-		}
-	}
-};
