+++ src/schema.graphql
@@ -14,7 +14,6 @@ type GetUserFeedResponse {
 
 type Query {
   GetUserFeed: GetUserFeedResponse!
-  hello: [Verification]
   GetCurrentUser: GetCurrentUserResponse!
 }
