+++ src/types/graph.d.ts
@@ -1,25 +1,14 @@
-export const typeDefs = ["type Comment {\n  id: String!\n  comment: String\n  user: User\n  feed: Feed\n  createAt: String!\n  updateAt: String!\n}\n\ntype GetUserFeedResponse {\n  res: Boolean!\n  error: String\n}\n\ntype Query {\n  GetUserFeed: GetUserFeedResponse!\n  hello: [Verification]\n  GetCurrentUser: GetCurrentUserResponse!\n}\n\ntype Feed {\n  id: String!\n  photo: String!\n  text: String\n  user: User\n  likes: [User]\n  comments: [Comment]\n  createAt: String!\n  updateAt: String!\n}\n\ntype EmailSignInResponse {\n  res: Boolean!\n  error: String\n  token: String\n}\n\ntype Mutation {\n  EmailSignIn(email: String!, password: String!): EmailSignInResponse!\n  EmailSignUp(firstName: String!, lastName: String!, password: String!, email: String!): EmailSignUpResponse!\n  RequestVerification(type: String!, payload: String!): RequestVerificationResponse!\n  ValidateVerification(payload: String!, key: String!): ValidateVerificationResponse!\n}\n\ntype EmailSignUpResponse {\n  res: Boolean!\n  error: String\n  token: String\n}\n\ntype GetCurrentUserResponse {\n  res: Boolean!\n  error: String\n  user: User\n}\n\ntype User {\n  id: String!\n  fbId: String\n  googleId: String\n  firstName: String!\n  lastName: String!\n  password: String!\n  profilePhoto: String!\n  email: String!\n  isEmailVerified: Boolean\n  phone: String!\n  isPhoneVerified: Boolean\n  feeds: [Feed]\n  likes: [Feed]\n  comments: [Comment]\n  createAt: String!\n  updateAt: String!\n}\n\ntype RequestVerificationResponse {\n  res: Boolean!\n  error: String\n}\n\ntype Verification {\n  id: String!\n  type: String!\n  payload: String!\n  key: String!\n  isVerified: Boolean!\n  createAt: String!\n  updateAt: String!\n}\n\ntype ValidateVerificationResponse {\n  res: Boolean!\n  error: String\n}\n"];
+export const typeDefs = ["type Comment {\n  id: String!\n  comment: String\n  user: User\n  feed: Feed\n  createAt: String!\n  updateAt: String!\n}\n\ntype GetUserFeedResponse {\n  res: Boolean!\n  error: String\n}\n\ntype Query {\n  GetUserFeed: GetUserFeedResponse!\n  GetCurrentUser: GetCurrentUserResponse!\n}\n\ntype Feed {\n  id: String!\n  photo: String!\n  text: String\n  user: User\n  likes: [User]\n  comments: [Comment]\n  createAt: String!\n  updateAt: String!\n}\n\ntype EmailSignInResponse {\n  res: Boolean!\n  error: String\n  token: String\n}\n\ntype Mutation {\n  EmailSignIn(email: String!, password: String!): EmailSignInResponse!\n  EmailSignUp(firstName: String!, lastName: String!, password: String!, email: String!): EmailSignUpResponse!\n  RequestVerification(type: String!, payload: String!): RequestVerificationResponse!\n  ValidateVerification(payload: String!, key: String!): ValidateVerificationResponse!\n}\n\ntype EmailSignUpResponse {\n  res: Boolean!\n  error: String\n  token: String\n}\n\ntype GetCurrentUserResponse {\n  res: Boolean!\n  error: String\n  user: User\n}\n\ntype User {\n  id: String!\n  fbId: String\n  googleId: String\n  firstName: String!\n  lastName: String!\n  password: String!\n  profilePhoto: String!\n  email: String!\n  isEmailVerified: Boolean\n  phone: String!\n  isPhoneVerified: Boolean\n  feeds: [Feed]\n  likes: [Feed]\n  comments: [Comment]\n  createAt: String!\n  updateAt: String!\n}\n\ntype RequestVerificationResponse {\n  res: Boolean!\n  error: String\n}\n\ntype Verification {\n  id: String!\n  type: String!\n  payload: String!\n  key: String!\n  isVerified: Boolean!\n  createAt: String!\n  updateAt: String!\n}\n\ntype ValidateVerificationResponse {\n  res: Boolean!\n  error: String\n}\n"];
 /* tslint:disable */
 
 export interface Query {
   GetUserFeed: GetUserFeedResponse;
-  hello: Array<Verification> | null;
   GetCurrentUser: GetCurrentUserResponse;
 }
 
 export interface GetUserFeedResponse {
   res: boolean;
   error: string | null;
-}
-
-export interface Verification {
-  id: string;
-  type: string;
-  payload: string;
-  key: string;
-  isVerified: boolean;
-  createAt: string;
-  updateAt: string;
 }
 
 export interface GetCurrentUserResponse {
@@ -116,4 +105,14 @@ export interface RequestVerificationResponse {
 export interface ValidateVerificationResponse {
   res: boolean;
   error: string | null;
+}
+
+export interface Verification {
+  id: string;
+  type: string;
+  payload: string;
+  key: string;
+  isVerified: boolean;
+  createAt: string;
+  updateAt: string;
 }
